FinanceApp/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.1.1
pytz==2023.3
pytest==7.4.3
httpx==0.25.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
email-validator==2.1.0
